move some tests to assetlibrarytest
