+++ composer.json
@@ -25,9 +25,8 @@
         "spatie/laravel-medialibrary": "^7.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~7.0 || ~8.0",
-        "orchestra/testbench": "^3.6",
-        "squizlabs/php_codesniffer": "^3.0"
+        "phpunit/phpunit": "~8.0",
+        "orchestra/testbench": "^3.6"
     },
     "autoload": {
         "psr-4": {
